requirements.txt
asyncio==3.4.3
playwright==1.41.1
bs4==0.0.2
pandas==2.0.3
tiktoken==0.0.1
faiss-cpu==1.8.0
transformers==4.41.2
torch==2.0.1
requests==2.31.0
python-dotenv==1.0.0

